feat(contracts): let verifier helper read unirep contract
