+++ packages/contracts/contracts/verifierHelpers/EpochKeyVerifierHelper.sol
@@ -1,11 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
+import {Unirep} from '../Unirep.sol';
 import {IVerifier} from '../interfaces/IVerifier.sol';
 import {BaseVerifierHelper} from './BaseVerifierHelper.sol';
 
 contract EpochKeyVerifierHelper is BaseVerifierHelper {
-    constructor(IVerifier _verifier) BaseVerifierHelper(_verifier) {}
+    constructor(
+        Unirep _unirep,
+        IVerifier _verifier
+    ) BaseVerifierHelper(_unirep, _verifier) {}
 
     function decodeEpochKeySignals(
         uint256[] calldata publicSignals
@@ -35,8 +39,18 @@ contract EpochKeyVerifierHelper is BaseVerifierHelper {
     ) public view returns (EpochKeySignals memory) {
         EpochKeySignals memory signals = decodeEpochKeySignals(publicSignals);
 
-        bool valid = verifier.verifyProof(publicSignals, proof);
-        if (!valid) revert InvalidProof();
+        if (!verifier.verifyProof(publicSignals, proof)) revert InvalidProof();
+
+        uint48 epoch = unirep.attesterCurrentEpoch(signals.attesterId);
+        if (signals.epoch > epoch) revert InvalidEpoch();
+
+        if (
+            !unirep.attesterStateTreeRootExists(
+                signals.attesterId,
+                signals.epoch,
+                signals.stateTreeRoot
+            )
+        ) revert InvalidStateTreeRoot(signals.stateTreeRoot);
 
         if (signals.chainId != chainid) revert ChainIdNotMatch(signals.chainId);
