+++ packages/contracts/contracts/verifierHelpers/ReputationVerifierHelper.sol
@@ -1,11 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
+import {Unirep} from '../Unirep.sol';
 import {IVerifier} from '../interfaces/IVerifier.sol';
 import {BaseVerifierHelper} from './BaseVerifierHelper.sol';
 
 contract ReputationVerifierHelper is BaseVerifierHelper {
-    constructor(IVerifier _verifier) BaseVerifierHelper(_verifier) {}
+    constructor(
+        Unirep _unirep,
+        IVerifier _verifier
+    ) BaseVerifierHelper(_unirep, _verifier) {}
 
     function decodeReputationSignals(
         uint256[] calldata publicSignals
@@ -83,8 +87,18 @@ contract ReputationVerifierHelper is BaseVerifierHelper {
             publicSignals
         );
 
-        bool valid = verifier.verifyProof(publicSignals, proof);
-        if (!valid) revert InvalidProof();
+        if (!verifier.verifyProof(publicSignals, proof)) revert InvalidProof();
+
+        uint48 epoch = unirep.attesterCurrentEpoch(signals.attesterId);
+        if (signals.epoch > epoch) revert InvalidEpoch();
+
+        if (
+            !unirep.attesterStateTreeRootExists(
+                signals.attesterId,
+                signals.epoch,
+                signals.stateTreeRoot
+            )
+        ) revert InvalidStateTreeRoot(signals.stateTreeRoot);
 
         if (signals.chainId != chainid) revert ChainIdNotMatch(signals.chainId);
